+++ yash-env/src/system/virtual.rs
@@ -928,11 +928,19 @@ impl System for VirtualSystem {
     /// The `execve` system call cannot be simulated in the userland. This
     /// function returns `ENOSYS` if the file at `path` is a native executable,
     /// `ENOEXEC` if a non-executable file, and `ENOENT` otherwise.
-    fn execve(&mut self, path: &CStr, args: &[CString], envs: &[CString]) -> Result<Infallible> {
+    fn execve(
+        &mut self,
+        path: &CStr,
+        args: &[CString],
+        envs: &[CString],
+    ) -> Pin<Box<dyn Future<Output = Result<Infallible>>>> {
         let os_path = UnixStr::from_bytes(path.to_bytes());
         let mut state = self.state.borrow_mut();
         let fs = &state.file_system;
-        let file = fs.get(os_path)?;
+        let file = match fs.get(os_path) {
+            Ok(file) => file,
+            Err(e) => return Box::pin(std::future::ready(Err(e))),
+        };
         // TODO Check file permissions
         let is_executable = matches!(
             &file.borrow().body,
@@ -949,9 +957,12 @@ impl System for VirtualSystem {
             let envs = envs.to_owned();
             process.last_exec = Some((path, args, envs));
 
-            Err(Errno::ENOSYS)
+            // TODO: We should abort the currently running task and start the new one.
+            // Just returning `pending()` would break existing tests that rely on
+            // the current behavior.
+            Box::pin(std::future::ready(Err(Errno::ENOSYS)))
         } else {
-            Err(Errno::ENOEXEC)
+            Box::pin(std::future::ready(Err(Errno::ENOEXEC)))
         }
     }
 
@@ -2253,7 +2264,7 @@ mod tests {
             Box::new(move |child_env| {
                 Box::pin(async move {
                     let path = CString::new(path).unwrap();
-                    let _ = child_env.system.execve(&path, &[], &[]);
+                    child_env.system.execve(&path, &[], &[]).await.ok();
                     child_env.system.exit(child_env.exit_status).await
                 })
             }),
@@ -2528,7 +2539,7 @@ mod tests {
         state.file_system.save(path, content).unwrap();
         drop(state);
         let path = CString::new(path).unwrap();
-        let result = system.execve(&path, &[], &[]);
+        let result = system.execve(&path, &[], &[]).now_or_never().unwrap();
         assert_eq!(result, Err(Errno::ENOSYS));
     }
 
@@ -2549,7 +2560,7 @@ mod tests {
         let path = CString::new(path).unwrap();
         let args = [c"file".to_owned(), c"bar".to_owned()];
         let envs = [c"foo=FOO".to_owned(), c"baz".to_owned()];
-        let _ = system.execve(&path, &args, &envs);
+        system.execve(&path, &args, &envs).now_or_never();
 
         let process = system.current_process();
         let arguments = process.last_exec.as_ref().unwrap();
@@ -2569,14 +2580,17 @@ mod tests {
         state.file_system.save(path, content).unwrap();
         drop(state);
         let path = CString::new(path).unwrap();
-        let result = system.execve(&path, &[], &[]);
+        let result = system.execve(&path, &[], &[]).now_or_never().unwrap();
         assert_eq!(result, Err(Errno::ENOEXEC));
     }
 
     #[test]
     fn execve_returns_enoent_on_file_not_found() {
         let mut system = VirtualSystem::new();
-        let result = system.execve(c"/no/such/file", &[], &[]);
+        let result = system
+            .execve(c"/no/such/file", &[], &[])
+            .now_or_never()
+            .unwrap();
         assert_eq!(result, Err(Errno::ENOENT));
     }